feat: Highest Salaries Report
Implement a report that shows all departments along with the highest salary within each department.

Create a new namespace for reports: App\Persistence\Reports, along with a Report interface that has only one method for now called generate.

Create the HighestSalaryPerDepartment report as an implementation of Report.

Add the corresponding Route, and a Controller for department reports.